+++ src/R3dImageRenderer.cc
@@ -18,14 +18,14 @@ namespace Aa
 // ImageRenderer::ClippingPlane ////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////
 
-    ImageRenderer::ClippingPlane::ClippingPlane (const dvec3 & p, const dvec3 & n) :
+    ImageRenderer::ClippingPlane::ClippingPlane (const vec3 & p, const vec3 & n) :
       active (true),
       point (p),
       normal (n)
     {
     }
 
-    bool ImageRenderer::ClippingPlane::test (const dvec3 & p) const
+    bool ImageRenderer::ClippingPlane::test (const vec3 & p) const
     {
       return DotProd (p - point, normal) >= 0;
     }
